third_party/blink/web_tests/html5lib/generated/run-tests10-data-expected.txt

CONSOLE WARNING: A descendant of a <select> does not follow the content model.
CONSOLE WARNING: A descendant of a <select> does not follow the content model.
../resources/tests10.dat:
4
5
17
18

Test 4 of 54 in ../resources/tests10.dat failed. Input:
<!DOCTYPE html><body><select><svg></svg></select>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <svg svg>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

Test 5 of 54 in ../resources/tests10.dat failed. Input:
<!DOCTYPE html><body><select><option><svg></svg></option></select>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         <svg svg>
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

Test 17 of 54 in ../resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><tr><td><select><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <svg svg>
|                 <svg g>
|                   "foo"
|                 <svg g>
|                   "bar"
|               <p>
|                 "baz"
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "foobarbaz"
|     <p>
|       "quux"

Test 18 of 54 in ../resources/tests10.dat failed. Input:
<!DOCTYPE html><body><table><select><svg><g>foo</g><g>bar</g><p>baz</table><p>quux
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <svg svg>
|         <svg g>
|           "foo"
|         <svg g>
|           "bar"
|       <p>
|         "baz"
|     <table>
|     <p>
|       "quux"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       "foobarbaz"
|     <table>
|     <p>
|       "quux"
